+++ inversion_tree.h
@@ -13,7 +13,7 @@ struct inversionNode {
     std::vector<std::shared_ptr<inversionNode>> m_children;
     struct matrix getInvertedMatrix(std::vector<int> & invalidIndices, int parent);
 
-    void insertInvertedMatrix(std::vector<int> &invalidIndices, struct matrix &matrix, size_t shards, int parent);
+    void insertInvertedMatrix(std::vector<int> &invalidIndices, struct matrix &matrix, int shards, int parent);
 };
 
 class inversionTree {
@@ -21,7 +21,7 @@ public:
     // newInversionTree initializes a tree for storing inverted matrices.
     // Note that the m_root node is the identity matrix as it implies
     // there were no errors with the original data.
-    static inversionTree newInversionTree(size_t dataShards, size_t parityShards);
+    static inversionTree newInversionTree(int dataShards, int parityShards);
 
     // GetInvertedMatrix returns the cached inverted matrix or nil if it
     // is not found in the tree keyed on the indices of invalid rows.
@@ -31,7 +31,7 @@ public:
     // keyed by the indices of invalid rows.  The total number of shards
     // is required for creating the proper length lists of child nodes for
     // each node.
-    int InsertInvertedMatrix(std::vector<int> & invalidIndices, matrix &matrix, size_t shards);
+    int InsertInvertedMatrix(std::vector<int> & invalidIndices, matrix &matrix, int shards);
 
 private:
     inversionNode m_root;
